// src/models/floreriaModel.ts
// floreriaModel.ts
export interface Floreria {
    Id_Floreria?: number;
    Id_Vendedor: number;
    Id_Inventario: number;
    Imagen: string;
    Nombre: string;
    Descripcion: string;
    Telefono: string;
    Correo_Electronico: string;
    Redes_Sociales: string;
    Direccion: string;
    Codigo_Postal: number;
    Municipio: string;
}